clash-panel/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyyaml==6.0.1
python-multipart==0.0.6
aiofiles==23.2.1 